# openapi-spec.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/games": {
            "get": {
                "operationId": "GamesController_getGames",
                "summary": "Get paginated games list with filtering",
                "parameters": [
                    {
                        "name": "sortOrder",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortBy",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "enum": [
                                "popular",
                                "newest",
                                "name",
                                "rtp"
                            ],
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "provider",
                        "required": false,
                        "in": "query",
                        "description": "Provider slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "required": false,
                        "in": "query",
                        "description": "Category slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Games"
                ]
            }
        },
...
        "/api/v1/games/category/{slug}": {
            "get": {
                "operationId": "GamesController_getGamesByCategory",
                "summary": "Get games by category",
                "parameters": [
                    {
                        "name": "slug",
                        "required": true,
                        "in": "path",
                        "description": "Category slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "minimum": 1,
                            "default": 1,
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "minimum": 1,
                            "maximum": 100,
                            "default": 20,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Games"
                ]
            }
        },
        "/api/v1/games/favorites": {
            "get": {
                "operationId": "GamesController_getUserFavorites",
                "summary": "Get user favorite games",
                "parameters": [
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "minimum": 1,
                            "default": 1,
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "minimum": 1,
                            "maximum": 100,
                            "default": 20,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Games"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/games/favorites/{gameId}": {
            "post": {
                "operationId": "GamesController_addToFavorites",
                "summary": "Add game to favorites",
                "parameters": [
                    {
                        "name": "gameId",
                        "required": true,
                        "in": "path",
                        "description": "Game ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Games"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            },
            "delete": {
                "operationId": "GamesController_removeFromFavorites",
                "summary": "Remove game from favorites",
                "parameters": [
                    {
                        "name": "gameId",
                        "required": true,
                        "in": "path",
                        "description": "Game ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Games"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/games/recent": {
            "get": {
                "operationId": "GamesController_getUserRecentGames",
                "summary": "Get user recently played games",
                "parameters": [
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "minimum": 1,
                            "default": 1,
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "minimum": 1,
                            "maximum": 100,
                            "default": 20,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Games"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/games/history": {
            "get": {
                "operationId": "GamesController_getGameHistory",
                "summary": "Get user game session history",
                "parameters": [
                    {
                        "name": "endDate",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "startDate",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Games"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/games/{slug}": {
            "get": {
                "operationId": "GamesController_getGameBySlug",
                "summary": "Get single game details by slug",
                "parameters": [
                    {
                        "name": "slug",
                        "required": true,
                        "in": "path",
                        "description": "Game slug",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Games"
                ]
            }
        },
        "/api/v1/games/{slug}/launch": {
            "post": {
                "operationId": "GamesController_launchGame",
                "summary": "Launch game and get session URL",
                "parameters": [
                    {
                        "name": "slug",
                        "required": true,
                        "in": "path",
                        "description": "Game slug",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "coinType": {
                                        "type": "string",
                                        "enum": [
                                            "GC",
                                            "SC"
                                        ],
                                        "description": "Coin type to play with"
                                    }
                                },
                                "required": [
                                    "coinType"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Games"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/games/{gameId}/is-favorited": {
            "get": {
                "operationId": "GamesController_isGameFavorited",
                "summary": "Check if game is in user favorites",
                "parameters": [
                    {
                        "name": "gameId",
                        "required": true,
                        "in": "path",
                        "description": "Game ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Games"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/games/{slug}/stats": {
            "get": {
                "operationId": "GamesController_getGameStats",
                "summary": "Get game statistics",
                "parameters": [
                    {
                        "name": "slug",
                        "required": true,
                        "in": "path",
                        "description": "Game slug",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Games"
                ]
            }
        },
...
        "/api/v1/providers/{slug}": {
            "get": {
                "operationId": "ProvidersController_getProviderBySlug",
                "summary": "Get provider details with their games",
                "parameters": [
                    {
                        "name": "slug",
                        "required": true,
                        "in": "path",
                        "description": "Provider slug",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Providers"
                ]
            }
        },
...
        "/api/v1/admin/games/{id}": {
            "get": {
                "operationId": "AdminController_getGame",
                "summary": "Get game details",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Admin"
                ],
                "security": [
                    {
                        "Admin-JWT": []
                    }
                ]
            },
            "patch": {
                "operationId": "AdminController_updateGame",
                "summary": "Update game",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Admin"
                ],
                "security": [
                    {
                        "Admin-JWT": []
                    }
                ]
            },
            "delete": {
                "operationId": "AdminController_deleteGame",
                "summary": "Soft delete game",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Admin"
                ],
                "security": [
                    {
                        "Admin-JWT": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "WBC 2026 API",
        "description": "WBC 2026 Social Casino Platform API Documentation",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "JWT-auth": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http",
                "name": "JWT",
                "description": "Enter JWT token",
                "in": "header"
            }
        },
...
    }
}